s2scrud/requirements.txt
asgiref==3.2.3
Django==3.0.2
pytz==2019.3
sqlparse==0.3.0
djangorestframework==3.11.0
python-dateutil==2.8.1
six==1.13.0
